backend/requirements.txt
aiohttp==3.6.2
aiohttp-security==0.4.0
aiohttp-session==2.7.0
aiopg==0.16.0
aioredis==1.3.0
aiofile==1.5.2
SQLAlchemy==1.3.0b3
passlib==1.7.1
PyYAML==5.1.2
aiosmtplib==1.1.0
alembic==1.0.7
boto3==1.10.6
